versioning: start modifying TF2BD_VERSION_* vars from now on
I didn't initally plan to make that much modifications to bot detector, so really didn't feel that it was necessary.
But now that I have 8 releases and possibly planning on more (depending on how much energy and time I have), this is kind of needed as people might be potentially confused on what version of my fork they have.

I've thought of using my own seperate version variable in case the main branch gets updated; but pazer has abandoned this software, so we're overwriting the existing variables instead. (less work for me)

in case it does get actually updated or something, the about pop up will show what exact version it is forked from.